libc_noux & 64bit: fix printf format-warning
Normally for intptr_t the printf format PRIiPTR should be used. As Genode
printf doens't support this format we cast it to long int (intptr_t is int
for 32 bit and long int for 64 bit).

ref #989

+++ ports/src/lib/libc_noux/plugin.cc
@@ -261,7 +261,7 @@ extern "C" uid_t geteuid()
 void *sbrk(intptr_t increment)
 {
 	if (verbose)
-		PDBG("not implemented %d", increment);
+		PDBG("not implemented %ld", (long int)increment);
 	errno = ENOMEM;
 	return reinterpret_cast<void *>(-1);
 }
